small improvements
- moving all test creator attached files to "test_creator.data" dir
- improve exiting
- catch JsonDecodeError when loading language
- move todo tasks to trello board

+++ test_creator/exit.py
@@ -1,4 +1,5 @@
 import os
+import signal
 import typing
 import dearpygui.dearpygui as dpg
 from settings import *
@@ -9,10 +10,19 @@ logger.setLevel(LOGGING_LEVEL)
 lockfile: typing.TextIO | None = None
 
 
-def exit_mionly():
-    logger.debug(f'Unlocking {TEST_CREATOR_LOCK_FILENAME}')
-    lockfile.close()
-    os.remove(TEST_CREATOR_LOCK_FILENAME)
+def stop_mionly(do_exit: bool = False):
+    if not lockfile.closed:
+        logger.debug(f'Unlocking {TEST_CREATOR_LOCK_FILENAME}')
+
+        lockfile.close()
+        os.remove(TEST_CREATOR_LOCK_FILENAME)
 
     logger.debug('Stopping DearPyGui')
     dpg.stop_dearpygui()
+
+    if do_exit:
+        exit_app()
+
+
+def exit_app():
+    os.kill(os.getpid(), signal.SIGINT)
